Fixed suricata.sh (#124)
It was syntax errors in the Suricata YAML configuration file that caused
the multiple
Initial service start failed, restart failed, etc...

The HOME_NET definition had an invalid format:
HOME_NET: "10.0.2.15/24
10.0.2.15/24"

was improperly formatted - it should be an array or a single value.

Interface Definition Errors: There were multiple malformed interface
definitions:
- interface: enp0s3 enp0s: 3
and
- interface: enp0s3 enp0s3

The errors  caused the YAML parser to fail when processing the config. 

I was able to automate a fix for these issues when running the script.

The script continued to run all the way through

Other than improving the YAML & script, it was necessacry to completely
uninstall and purge suricata
before retrying.

Maybe Give it a go on your VM just to make sure.

---------

Signed-off-by: Chris <atnl@protonmail.com>
Signed-off-by: Tim Burns <92951498+OpenSource-For-Freedom@users.noreply.github.com>
Co-authored-by: LinuxUser255 <atnl@protonmailcom>
Co-authored-by: Tim Burns <92951498+OpenSource-For-Freedom@users.noreply.github.com>